+++ Data/Reagents.lua
@@ -2,6 +2,7 @@ TF_REAGENT_VENDORS = {
     { zone = TF_ZONE.ORGRIMMAR, x = 0.4574, y = 0.4094, name = "Magenius", faction = TF_FACTION.HORDE },
     { zone = TF_ZONE.ORGRIMMAR, x = 0.4545, y = 0.5654, name = "Horthus", faction = TF_FACTION.HORDE },
     { zone = TF_ZONE.ORGRIMMAR, x = 0.4599, y = 0.4569, name = "Hagrus", faction = TF_FACTION.HORDE },
+    { zone = TF_ZONE.ASHENVALE, x = 0.3498, y = 0.4846, name = "Llana", faction = TF_FACTION.ALLIANCE },
     { zone = TF_ZONE.THE_BARRENS, x = 0.5139, y = 0.3020, name = "Hula'mahi", faction = TF_FACTION.HORDE },
     { zone = TF_ZONE.STONETALON_MOUNTAINS, x = 0.4762, y = 0.6159, name = "Jeeda", faction = TF_FACTION.HORDE },
     { zone = TF_ZONE.UNDERCITY, x = 0.6970, y = 0.3906, name = "Thomas Mordan", faction = TF_FACTION.HORDE },