+++ changelog.md
@@ -2,7 +2,11 @@
 * Added Darkshore NPCs
 * Added Redridge Mountains NPCs
 * Added Duskwood NPCs
+* Added Alliance Ashenvale NPCs
+* Added Alliance Azshara NPCs
+* Added Theramore Isle, Dustwallow Marsh NPCs
 * Added several missing NPCs to Stranglethorn Vale
+* Added several missing NPCs to Wetlands
 * Minor NPC fixes in Stormwind, Darnassus
 
 [See past changes here.](https://bitbucket.org/jsiebert9/townsfolk-tracker/src/master/changehistory.md)